fix(build): add missing OpenJTalk library files to BUILD_BYPRODUCTS
The build was failing on macOS ARM64 because libnjd_set_pronunciation.a and other
njd_set_* libraries were not listed in BUILD_BYPRODUCTS. This caused CMake's ninja
generator to not know these files would be created by the ExternalProject, leading
to the error: 'oj/lib/libnjd_set_pronunciation.a', needed by 'piper', missing and
no known rule to make it.

Added all the njd_set_* libraries that are copied in the INSTALL_COMMAND to the
BUILD_BYPRODUCTS list to ensure proper dependency tracking.